feat(models): reorganizar la estructura de modelos con responsabilidad única
- Separa cada modelo en archivos independientes (, ,
  , ).
- Reemplaza ifs repetitivos por métodos descriptivos (e.g. , ,
  ).
- Ajusta las importaciones en  y el resto del código (services, steps de Behave).
- Verifica que los tests BDD (Behave) se ejecutan correctamente.
