--- .idea/inspectionProfiles/Project_Default.xml
@@ -1,6 +1,0 @@
-<component name="InspectionProjectProfileManager">
-  <profile version="1.0">
-    <option name="myName" value="Project Default" />
-    <inspection_tool class="CucumberPlusUndefinedStep" enabled="false" level="WARNING" enabled_by_default="false" />
-  </profile>
-</component>
