+++ ciudadano_app/models/__init__.py
@@ -1,0 +1,7 @@
+"""
+Inicializa el paquete de modelos para la aplicación ciudadano_app.
+"""
+from .ciudadano import Ciudadano
+from .gestor_ciudadano import GestorCiudadano
+
+__all__ = ['Ciudadano', 'GestorCiudadano']
