+++ entidad_municipal_app/models/__init__.py
@@ -1,0 +1,7 @@
+"""
+Inicializa el paquete de modelos para la aplicación entidad_municipal_app.
+"""
+from .evento_municipal import EventoMunicipal
+from .registro_asistencia import RegistroAsistencia
+
+__all__ = ['EventoMunicipal', 'RegistroAsistencia']
